+++ src/main/java/framework/base/BrowserType.java
@@ -3,6 +3,7 @@ package framework.base;
 public enum BrowserType {
     firefox,
     chrome,
+    chrome_headless,
     ie,
     opera,
     edge